Move common RSA util functinos to separate file
