+++ src/soter/openssl/soter_rsa_key_3.c
@@ -28,104 +28,8 @@
 #include <openssl/param_build.h>
 #include <openssl/rsa.h>
 
+#include "soter/openssl/soter_rsa_key_utils.h"
 #include "soter/soter_portable_endian.h"
-
-static size_t rsa_pub_key_size(int mod_size)
-{
-    switch (mod_size) {
-    case 128: /* 1024 */
-        return sizeof(soter_rsa_pub_key_1024_t);
-    case 256: /* 2048 */
-        return sizeof(soter_rsa_pub_key_2048_t);
-    case 512: /* 4096 */
-        return sizeof(soter_rsa_pub_key_4096_t);
-    case 1024: /* 8192 */
-        return sizeof(soter_rsa_pub_key_8192_t);
-    default:
-        return 0;
-    }
-}
-
-static size_t rsa_priv_key_size(int mod_size)
-{
-    switch (mod_size) {
-    case 128: /* 1024 */
-        return sizeof(soter_rsa_priv_key_1024_t);
-    case 256: /* 2048 */
-        return sizeof(soter_rsa_priv_key_2048_t);
-    case 512: /* 4096 */
-        return sizeof(soter_rsa_priv_key_4096_t);
-    case 1024: /* 8192 */
-        return sizeof(soter_rsa_priv_key_8192_t);
-    default:
-        return 0;
-    }
-}
-
-static char* rsa_pub_key_tag(int mod_size)
-{
-    switch (mod_size) {
-    case 128: /* 1024 */
-        return RSA_PUB_KEY_TAG(1024);
-    case 256: /* 2048 */
-        return RSA_PUB_KEY_TAG(2048);
-    case 512: /* 4096 */
-        return RSA_PUB_KEY_TAG(4096);
-    case 1024: /* 8192 */
-        return RSA_PUB_KEY_TAG(8192);
-    default:
-        return NULL;
-    }
-}
-
-static char* rsa_priv_key_tag(int mod_size)
-{
-    switch (mod_size) {
-    case 128: /* 1024 */
-        return RSA_PRIV_KEY_TAG(1024);
-    case 256: /* 2048 */
-        return RSA_PRIV_KEY_TAG(2048);
-    case 512: /* 4096 */
-        return RSA_PRIV_KEY_TAG(4096);
-    case 1024: /* 8192 */
-        return RSA_PRIV_KEY_TAG(8192);
-    default:
-        return NULL;
-    }
-}
-
-static bool is_mod_size_supported(int mod_size)
-{
-    switch (mod_size) {
-    case 128:
-    case 256:
-    case 512:
-    case 1024:
-        return true;
-    default:
-        return false;
-    }
-}
-
-static soter_status_t bignum_to_bytes(const BIGNUM* bn, uint8_t* to, size_t to_length)
-{
-    size_t bn_size = (size_t)BN_num_bytes(bn);
-    size_t bytes_copied;
-
-    if (bn_size > to_length) {
-        return SOTER_FAIL;
-    }
-
-    bytes_copied = BN_bn2bin(bn, to + (to_length - bn_size));
-
-    if (bytes_copied != bn_size) {
-        return SOTER_FAIL;
-    }
-
-    memset(to, 0, to_length - bn_size);
-
-    return SOTER_SUCCESS;
-}
 
 soter_status_t soter_engine_specific_to_rsa_pub_key(const soter_engine_specific_rsa_key_t* engine_key,
                                                     soter_container_hdr_t* key,
